// src/redux/tag-types.ts
export enum TagTypes {
  profile = "profile",
  contact = "contact",
  skill = "skill",
  skill_category = "skill_category",
  project = "project",
  blog = "blog",
  education = "education",
  experience = "experience",
  certification = "certification",
}

export const tagTypesList = [
  TagTypes.profile,
  TagTypes.contact,
  TagTypes.skill,
  TagTypes.skill_category,
  TagTypes.project,
  TagTypes.blog,
  TagTypes.education,
  TagTypes.experience,
  TagTypes.certification,
];